Add OCI image annotations and use annotations (not labels)
These annotations are useful for tools (such as Renovate and Snyk) to use as well as for manual use by individuals.

The specification says to use annotations in place of labels when
possible so migrate to annotations.

Add version and revision annotations.

See: https://github.com/opencontainers/image-spec/blob/v1.1.0/annotations.md#pre-defined-annotation-keys